requirements.txt
absl-py==1.0.0
astunparse==1.6.3
autocorrect==2.6.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
backcall==0.2.0
beautifulsoup4==4.10.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
cycler==0.11.0
decorator==5.1.0
docformatter==1.4
Flask==2.0.2
flask-ngrok==0.0.25
flatbuffers==2.0
fonttools==4.28.5
gast==0.5.3
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
gunicorn==20.1.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.10.0
ipython==7.30.0
isodate==0.6.1
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==13.0.0
lxml==4.6.4
Mako==1.1.6
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
msrest==0.6.21
nltk==3.6.5
numpy==1.22.0
oauthlib==3.1.1
opencv-contrib-python==4.5.5.62
opencv-python==4.5.5.62
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
parso==0.8.2
pdoc==8.0.1
pdoc3==0.10.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
prompt-toolkit==3.0.23
protobuf==3.19.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
Pygments==2.10.0
pyparsing==3.0.6
python-aiml==0.9.3
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
regex==2021.11.10
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
scikit-learn==1.0.1
scipy==1.7.3
seaborn==0.11.2
six==1.16.0
sklearn==0.0
soupsieve==2.3.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
thop==0.0.31.post2005241907
threadpoolctl==3.0.0
torch==1.10.1
torchvision==0.11.2
tqdm==4.62.3
traitlets==5.1.1
typing-extensions==4.0.1
untokenize==0.1.1
urllib3==1.26.8
wcwidth==0.2.5
Werkzeug==2.0.2
wikipedia==1.4.0
wrapt==1.14.0
yapf==0.32.0
zipp==3.7.0
